fetchCheriSDK.groovy: Use dev compiler by default
Previously only a few named branches of CheriBSD (including dev) get built by the dev compiler. This creates a situation where CheriBSD PRs are built by the main compiler but merged into the dev branch, which is built by the dev compiler.

This commit changes the default to use the dev compiler and only the main branch uses the master compiler.