vars/cheribuildProject.groovy: Attempt to fix long-standing status race
PullRequestGlobalVariable appears to not save the new object in the
script's bindings like other subclasses of GlobalVariable, and so I am
assuming that the racing we see is because every reference gets a fresh
instance, and it is during the constructor that the PR's metadata,
including the head commit, is retrieved. Even if it did, there would
still technically be a slight race between the build being triggered for
a specific commit and the first time pullRequest is referenced, so
although that's a bug in the plugin, we should still handle the case
where the pullRequest object does not correspond to what we're building.

This is all rather grotty due to GIT_COMMIT being the result of merging
the PR into the target branch (which may be a fast-forward, so we can't
always just take the second parent), with no equivalent for the
underlying PR commit we're testing.